requirements.txt
openai>=1.35.10
streamlit>=1.37.0
plotly>=5.24.0
streamlit-mic-recorder
python-dotenv
openai>=1.40.0
anthropic>=0.34.0
